feat(volunteering): reorganize project structure and implement MongoDB connection
